refactor(wizard): extract 7 step components into wizard/Step*.tsx
Phase 6 final slice. Each onboarding step is now its own file in
wizard/. Wizard.tsx becomes a thin orchestrator (state + nav + glue).

  - web/src/components/onboarding/Wizard.tsx: -1116 LOC (1981 -> 865)
  - wizard/Step1Welcome.tsx: 28 LOC (new)
  - wizard/Step2Templates.tsx: 139 LOC (new)
  - wizard/Step3Identity.tsx: 122 LOC (new)
  - wizard/Step4Team.tsx: 83 LOC (new)
  - wizard/Step5Setup.tsx: 482 LOC (new) — biggest; full runtime + memory grid
  - wizard/Step6Task.tsx: 109 LOC (new)
  - wizard/Step7Ready.tsx: 108 LOC (new)
  - wizard/runtime-helpers.ts: 38 LOC (new) — runtimeIsReady + detectedBinary

Wizard.tsx falls below the 1500-LOC fail threshold (now in 800-1500
warn zone). Allowlist entry removed.

Cumulative Phase 6:
  2641 -> 2412 (types + constants)         -229
       -> 1981 (helper components)         -431
       -> 865  (step components + helpers) -1116
  Net: -1776 LOC; Wizard.tsx shrunk by 67%.

No behavior change. Typecheck clean. All 565 web tests pass.
Build clean.

Step5Setup.tsx is over the 800 LOC warn threshold but well under 1500;
its 482-LOC bulk reflects the actual complexity of the runtime grid +
memory backend + Nex/GBrain key panels. Splitting further would
fragment the canContinue gate that ties them all together.
