+++ web/src/components/onboarding/Wizard.tsx
@@ -11,1151 +11,41 @@ import { useAppStore } from "../../stores/app";
 import { Kbd, MOD_KEY } from "../ui/Kbd";
 import "../../styles/onboarding.css";
 
-import { ApiKeyRow } from "./wizard/ApiKeyRow";
-
 // ApiKeyRow is re-exported here for back-compat with onboarding/ApiKeyRow.test.tsx.
-export { ApiKeyRow };
+export { ApiKeyRow } from "./wizard/ApiKeyRow";
 
+import { ProgressDots } from "./wizard/components";
 import {
-  ArrowIcon,
-  CheckIcon,
-  EnterHint,
-  ProgressDots,
-} from "./wizard/components";
-import {
-  API_KEY_FIELDS,
-  BLUEPRINT_CATEGORIES,
-  BLUEPRINT_DISPLAY,
   LOCAL_PROVIDER_LABELS,
   MEMORY_BACKEND_OPTIONS,
   RUNTIMES,
   SCRATCH_FOUNDING_TEAM,
   STEP_ORDER,
 } from "./wizard/constants";
-import { LocalLLMPicker } from "./wizard/LocalLLMPicker";
-import { NexSignupPanel } from "./wizard/NexSignupPanel";
+import { detectedBinary, runtimeIsReady } from "./wizard/runtime-helpers";
+import { WelcomeStep } from "./wizard/Step1Welcome";
+import { TemplatesStep } from "./wizard/Step2Templates";
+import { IdentityStep } from "./wizard/Step3Identity";
+import { TeamStep } from "./wizard/Step4Team";
+import { SetupStep } from "./wizard/Step5Setup";
+import { TaskStep } from "./wizard/Step6Task";
+import { ReadyStep } from "./wizard/Step7Ready";
 import type {
   BlueprintAgent,
-  BlueprintCategoryKey,
-  BlueprintDisplay,
   BlueprintTemplate,
   MemoryBackend,
   NexSignupStatus,
   PrereqResult,
   ReadinessCheck,
   ReadinessStatus,
-  RuntimeSpec,
   TaskTemplate,
   WizardStep,
 } from "./wizard/types";
 
-// runtimeIsReady centralizes the "should this runtime label count as a
-// configured LLM?" predicate used at the SetupStep gate, the keyboard
-// gate, and the ReadyStep summary. Three call sites had drifted apart
-// once already (see PR #367 review) — keeping the rule in one place
-// stops the next drift in its tracks.
-//
-// Rules:
-//   - Unknown labels (not in RUNTIMES) never count.
-//   - Runtimes with provider:null (Cursor/Windsurf) NEVER count, in
-//     either branch. finishOnboarding silently drops them from
-//     providerPriority, so a Cursor-only selection would let the gate
-//     pass and /config would persist no llm_provider.
-//   - With a non-null provider AND prereqs detection succeeded, the
-//     runtime's binary must be on PATH (detection.found).
-//   - With a non-null provider AND prereqs detection FAILED
-//     (prereqsError truthy), trust the user's selection.
-function runtimeIsReady(
-  label: string,
-  prereqs: PrereqResult[],
-  prereqsError: string,
-): boolean {
-  const spec = RUNTIMES.find((r) => r.label === label);
-  if (!spec || spec.provider === null) return false;
-  if (prereqsError) return true;
-  return Boolean(detectedBinary(prereqs, spec.binary)?.found);
-}
-
-// "Start from scratch" starter roster. Mirrors scratchFoundingTeamBlueprint
-// in internal/team/broker_onboarding.go — the broker seeds these exact slugs
-// when the wizard POSTs blueprint:null. Kept in sync manually; backend is the
-// source of truth, this is just the Team-step preview so users don't see an
-// empty roster before confirming.
-
+// runtimeIsReady + detectedBinary moved to wizard/runtime-helpers.ts.
 // ArrowIcon, CheckIcon, EnterHint, ProgressDots moved to wizard/components.tsx.
-
-/* ─── Step 1: Welcome ─── */
-
-interface WelcomeStepProps {
-  onNext: () => void;
-}
-
-function WelcomeStep({ onNext }: WelcomeStepProps) {
-  return (
-    <div className="wizard-step">
-      <div className="wizard-hero">
-        <div className="wizard-eyebrow">
-          <span className="status-dot active pulse" />
-          Ready to set up
-        </div>
-        <h1 className="wizard-headline">{ONBOARDING_COPY.step1_headline}</h1>
-        <p className="wizard-subhead">{ONBOARDING_COPY.step1_subhead}</p>
-      </div>
-      <div style={{ display: "flex", justifyContent: "center" }}>
-        <button className="btn btn-primary" onClick={onNext}>
-          {ONBOARDING_COPY.step1_cta}
-          <ArrowIcon />
-          <EnterHint />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-/* ─── Step 2: Templates ─── */
-
-interface TemplatesStepProps {
-  templates: BlueprintTemplate[];
-  loading: boolean;
-  selected: string | null;
-  onSelect: (id: string | null) => void;
-  onNext: () => void;
-  onBack: () => void;
-}
-
-function TemplatesStep({
-  templates,
-  loading,
-  selected,
-  onSelect,
-  onNext,
-  onBack,
-}: TemplatesStepProps) {
-  // Group templates by display category. Unknown blueprint ids (not in the
-  // frontend catalog) land in a catch-all "Other" bucket so new backend
-  // templates still render, just without the short-description and icon
-  // treatment.
-  const grouped = new Map<
-    BlueprintCategoryKey | "other",
-    BlueprintTemplate[]
-  >();
-  for (const t of templates) {
-    const display = BLUEPRINT_DISPLAY[t.id];
-    const key: BlueprintCategoryKey | "other" = display?.category ?? "other";
-    const list = grouped.get(key) ?? [];
-    list.push(t);
-    grouped.set(key, list);
-  }
-
-  const renderTile = (t: BlueprintTemplate) => {
-    const display = BLUEPRINT_DISPLAY[t.id];
-    const icon = display?.icon ?? t.emoji;
-    const desc = display?.shortDescription ?? t.description;
-    return (
-      <button
-        key={t.id}
-        className={`template-card ${selected === t.id ? "selected" : ""}`}
-        onClick={() => onSelect(t.id)}
-        type="button"
-      >
-        {icon && <div className="template-card-emoji">{icon}</div>}
-        <div className="template-card-name">{t.name}</div>
-        <div className="template-card-desc">{desc}</div>
-      </button>
-    );
-  };
-
-  return (
-    <div className="wizard-step">
-      <div className="wizard-hero">
-        <div className="wizard-eyebrow">
-          <span className="status-dot active pulse" />
-          Start with a preset, or build from scratch
-        </div>
-        <h1 className="wizard-headline">What should your office run?</h1>
-        <p className="wizard-subhead">
-          Pick the shape of work. We&apos;ll assemble the team, channels, and
-          first tasks around it. You can change anything later.
-        </p>
-      </div>
-
-      {loading ? (
-        <div className="wizard-panel">
-          <div
-            style={{
-              color: "var(--text-tertiary)",
-              fontSize: 13,
-              textAlign: "center",
-              padding: 20,
-            }}
-          >
-            Loading blueprints&hellip;
-          </div>
-        </div>
-      ) : (
-        <>
-          {BLUEPRINT_CATEGORIES.map((cat) => {
-            const items = grouped.get(cat.key) ?? [];
-            if (items.length === 0) return null;
-            return (
-              <div key={cat.key} className="wizard-panel template-group">
-                <div className="template-group-head">
-                  <p className="template-group-label">{cat.label}</p>
-                  <p className="template-group-hint">{cat.hint}</p>
-                </div>
-                <div className="template-grid">{items.map(renderTile)}</div>
-              </div>
-            );
-          })}
-
-          {(grouped.get("other") ?? []).length > 0 && (
-            <div className="wizard-panel template-group">
-              <div className="template-group-head">
-                <p className="template-group-label">Other</p>
-              </div>
-              <div className="template-grid">
-                {(grouped.get("other") ?? []).map(renderTile)}
-              </div>
-            </div>
-          )}
-
-          <div className="template-from-scratch">
-            <button
-              className={`template-from-scratch-btn ${selected === null ? "selected" : ""}`}
-              onClick={() => onSelect(null)}
-              type="button"
-            >
-              <span className="template-from-scratch-icon">+</span>
-              Start from scratch
-              <span className="template-from-scratch-sub">
-                5-person founding team: CEO, GTM Lead, Founding Engineer, PM,
-                Designer
-              </span>
-            </button>
-          </div>
-        </>
-      )}
-
-      <div className="wizard-nav">
-        <button className="btn btn-ghost" onClick={onBack} type="button">
-          Back
-        </button>
-        <button className="btn btn-primary" onClick={onNext} type="button">
-          Review the team
-          <ArrowIcon />
-          <EnterHint />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-/* ─── Step 3: Identity ─── */
-
-// NexSignupStatus tracks the state of the optional in-wizard Nex
-// registration sub-flow. 'hidden' means the user hasn't opened the
-// affordance yet; 'open' means they're entering their email; 'submitting'
-// is the in-flight POST to /nex/register; 'ok' shows a green "sent, check
-// your inbox" hint; 'fallback' flips to the external-link version when
-// nex-cli is not installed (the broker responds 502 with ErrNotInstalled).
-
-interface IdentityStepProps {
-  company: string;
-  description: string;
-  priority: string;
-  nexEmail: string;
-  nexSignupStatus: NexSignupStatus;
-  nexSignupError: string;
-  onChangeCompany: (v: string) => void;
-  onChangeDescription: (v: string) => void;
-  onChangePriority: (v: string) => void;
-  onChangeNexEmail: (v: string) => void;
-  onSubmitNexSignup: () => void;
-  onOpenNexSignup: () => void;
-  onNext: () => void;
-  onBack: () => void;
-}
-
-function IdentityStep({
-  company,
-  description,
-  priority,
-  nexEmail,
-  nexSignupStatus,
-  nexSignupError,
-  onChangeCompany,
-  onChangeDescription,
-  onChangePriority,
-  onChangeNexEmail,
-  onSubmitNexSignup,
-  onOpenNexSignup,
-  onNext,
-  onBack,
-}: IdentityStepProps) {
-  const canContinue =
-    company.trim().length > 0 && description.trim().length > 0;
-
-  return (
-    <div className="wizard-step">
-      <div className="wizard-panel">
-        <p className="wizard-panel-title">Tell us about this office</p>
-        <div className="form-group">
-          <label className="label" htmlFor="wiz-company">
-            Company or project name{" "}
-            <span style={{ color: "var(--red)" }}>*</span>
-          </label>
-          <input
-            className="input"
-            id="wiz-company"
-            placeholder="Acme Operations, or your real project name"
-            autoComplete="organization"
-            value={company}
-            onChange={(e) => onChangeCompany(e.target.value)}
-          />
-        </div>
-        <div className="form-group">
-          <label className="label" htmlFor="wiz-description">
-            One-liner description <span style={{ color: "var(--red)" }}>*</span>
-          </label>
-          <input
-            className="input"
-            id="wiz-description"
-            placeholder="What real business or workflow should this office run?"
-            value={description}
-            onChange={(e) => onChangeDescription(e.target.value)}
-          />
-        </div>
-        <div className="form-group">
-          <label className="label" htmlFor="wiz-priority">
-            Top priority right now
-          </label>
-          <input
-            className="input"
-            id="wiz-priority"
-            placeholder="Win the first real customer loop"
-            value={priority}
-            onChange={(e) => onChangePriority(e.target.value)}
-          />
-        </div>
-      </div>
-
-      {nexSignupStatus === "hidden" ? (
-        <div className="wiz-nex-trigger">
-          <button
-            type="button"
-            className="wiz-nex-trigger-link"
-            onClick={onOpenNexSignup}
-          >
-            Don&apos;t have a Nex account? Sign up here.
-          </button>
-        </div>
-      ) : (
-        <NexSignupPanel
-          email={nexEmail}
-          status={nexSignupStatus}
-          error={nexSignupError}
-          onChangeEmail={onChangeNexEmail}
-          onSubmit={onSubmitNexSignup}
-        />
-      )}
-
-      <div className="wizard-nav">
-        <button className="btn btn-ghost" onClick={onBack} type="button">
-          Back
-        </button>
-        <button
-          className="btn btn-primary"
-          onClick={onNext}
-          disabled={!canContinue}
-          type="button"
-        >
-          Choose a blueprint
-          <ArrowIcon />
-          <EnterHint />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-/* ─── Step 4: Team Review ─── */
-
-interface TeamStepProps {
-  agents: BlueprintAgent[];
-  onToggle: (slug: string) => void;
-  onNext: () => void;
-  onBack: () => void;
-}
-
-function TeamStep({ agents, onToggle, onNext, onBack }: TeamStepProps) {
-  return (
-    <div className="wizard-step">
-      <div className="wizard-panel">
-        <p className="wizard-panel-title">Your team</p>
-        <p
-          style={{
-            fontSize: 12,
-            color: "var(--text-secondary)",
-            margin: "-8px 0 12px 0",
-          }}
-        >
-          These are the specialists your blueprint assembled. Toggle anyone you
-          don&apos;t need.
-        </p>
-
-        {agents.length === 0 ? (
-          <div className="wiz-team-empty">
-            No teammates yet. Go back and pick a blueprint, or open the office
-            and add agents from the team panel.
-          </div>
-        ) : (
-          <div className="wiz-team-grid">
-            {agents.map((a) => {
-              // Lead agent is always included and cannot be unchecked here.
-              // The backend also refuses to remove or disable any BuiltIn
-              // member, so this is UI belt + server-side braces.
-              const locked = a.built_in === true;
-              return (
-                <button
-                  key={a.slug}
-                  className={`wiz-team-tile ${a.checked ? "selected" : ""} ${locked ? "locked" : ""}`}
-                  onClick={() => !locked && onToggle(a.slug)}
-                  type="button"
-                  disabled={locked}
-                  aria-disabled={locked}
-                  title={locked ? "Lead agent — always included" : undefined}
-                >
-                  <div className="wiz-team-check">
-                    {a.checked && <CheckIcon />}
-                  </div>
-                  <div>
-                    {a.emoji && (
-                      <span style={{ marginRight: 6 }}>{a.emoji}</span>
-                    )}
-                    <span className="wiz-team-name">{a.name}</span>
-                    {locked && (
-                      <span className="wiz-team-lead-badge" aria-label="Lead">
-                        Lead
-                      </span>
-                    )}
-                    {a.role && <div className="wiz-team-role">{a.role}</div>}
-                  </div>
-                </button>
-              );
-            })}
-          </div>
-        )}
-      </div>
-
-      <div className="wizard-nav">
-        <button className="btn btn-ghost" onClick={onBack} type="button">
-          Back
-        </button>
-        <button className="btn btn-primary" onClick={onNext} type="button">
-          Continue
-          <ArrowIcon />
-          <EnterHint />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-/* ─── Step 5: Setup ─── */
-
-interface SetupStepProps {
-  prereqs: PrereqResult[];
-  prereqsLoading: boolean;
-  prereqsError: string;
-  runtimePriority: string[];
-  onToggleRuntime: (label: string) => void;
-  onReorderRuntime: (label: string, direction: -1 | 1) => void;
-  apiKeys: Record<string, string>;
-  onChangeApiKey: (key: string, value: string) => void;
-  memoryBackend: MemoryBackend;
-  onChangeMemoryBackend: (value: MemoryBackend) => void;
-  nexApiKey: string;
-  onChangeNexApiKey: (v: string) => void;
-  gbrainOpenAIKey: string;
-  onChangeGBrainOpenAIKey: (v: string) => void;
-  gbrainAnthropicKey: string;
-  onChangeGBrainAnthropicKey: (v: string) => void;
-  // Local-LLM opt-in chosen here; submitted with the rest of the wizard
-  // payload at /onboarding/complete and applied to llm_provider so the
-  // user's selection takes effect on first agent turn.
-  localProvider: string;
-  onSelectLocalProvider: (kind: string) => void;
-  onNext: () => void;
-  onBack: () => void;
-}
-
-function detectedBinary(
-  prereqs: PrereqResult[],
-  binary: string,
-): PrereqResult | undefined {
-  return prereqs.find((p) => p.name === binary);
-}
-
-function SetupStep({
-  prereqs,
-  prereqsLoading,
-  prereqsError,
-  runtimePriority,
-  onToggleRuntime,
-  onReorderRuntime,
-  apiKeys,
-  onChangeApiKey,
-  memoryBackend,
-  onChangeMemoryBackend,
-  nexApiKey,
-  onChangeNexApiKey,
-  gbrainOpenAIKey,
-  onChangeGBrainOpenAIKey,
-  gbrainAnthropicKey,
-  onChangeGBrainAnthropicKey,
-  localProvider,
-  onSelectLocalProvider,
-  onNext,
-  onBack,
-}: SetupStepProps) {
-  // localModeOn governs whether the second-step LocalLLMPicker is
-  // shown beneath the runtime grid. Initialised from the parent's
-  // localProvider so re-entering the step preserves the user's
-  // earlier "I want a local model" choice. Toggling the meta-tile
-  // off clears localProvider via onSelectLocalProvider("").
-  const [localModeOn, setLocalModeOn] = useState<boolean>(localProvider !== "");
-
-  // Any priority slot that satisfies runtimeIsReady satisfies the gate.
-  const hasInstalledSelection = runtimePriority.some((label) =>
-    runtimeIsReady(label, prereqs, prereqsError),
-  );
-  const hasAnyApiKey = Object.values(apiKeys).some((v) => v.trim().length > 0);
-  // GBrain requires an OpenAI key to function — the TUI gates on this in
-  // InitGBrainOpenAIKey (see internal/tui/init_flow.go:215). Mirror the
-  // gate here so the wizard doesn't let users commit an unusable config.
-  const gbrainSelected = memoryBackend === "gbrain";
-  const gbrainOpenAIMissing =
-    gbrainSelected && gbrainOpenAIKey.trim().length === 0;
-  // Picking a local LLM is an alternative path — once selected the user
-  // doesn't need a cloud CLI installed or any cloud API key set.
-  const hasLocalProvider = localProvider.trim().length > 0;
-  const canContinue =
-    (hasInstalledSelection || hasAnyApiKey || hasLocalProvider) &&
-    !gbrainOpenAIMissing;
-
-  return (
-    <div className="wizard-step">
-      <div className="wizard-panel">
-        <p className="wizard-panel-title">How should agents run?</p>
-        <p
-          style={{
-            fontSize: 12,
-            color: "var(--text-secondary)",
-            margin: "-8px 0 12px 0",
-          }}
-        >
-          Pick the CLIs you have installed. Each CLI&apos;s login handles its
-          own provider auth, so no API keys are needed. Select multiple to set a
-          fallback order — if the first one fails, agents fall through to the
-          next.
-        </p>
-
-        {prereqsLoading ? (
-          <div
-            style={{
-              color: "var(--text-tertiary)",
-              fontSize: 13,
-              padding: "8px 0",
-            }}
-          >
-            Checking which CLIs are installed&hellip;
-          </div>
-        ) : prereqsError ? (
-          <div
-            data-testid="prereqs-error-banner"
-            role="alert"
-            style={{
-              fontSize: 12,
-              color: "var(--danger-500, #c33)",
-              padding: "10px 12px",
-              background: "var(--danger-50, #fee)",
-              border: "1px solid var(--danger-200, #fcc)",
-              borderRadius: 6,
-              marginBottom: 12,
-            }}
-          >
-            Could not detect installed CLIs:{" "}
-            <code style={{ fontFamily: "var(--font-mono)" }}>
-              {prereqsError}
-            </code>
-            . You can still select a runtime below or add an API key to
-            continue.
-          </div>
-        ) : null}
-        {prereqsLoading ? null : (
-          <div className="runtime-grid">
-            {RUNTIMES.map((spec) => {
-              const detection = detectedBinary(prereqs, spec.binary);
-              const installed = Boolean(detection?.found);
-              const selectable = installed || Boolean(prereqsError);
-              const priorityIdx = runtimePriority.indexOf(spec.label);
-              const selected = priorityIdx >= 0;
-              const classes = [
-                "runtime-tile",
-                selected ? "selected" : "",
-                selectable ? "" : "disabled",
-              ]
-                .filter(Boolean)
-                .join(" ");
-              return (
-                <button
-                  key={spec.label}
-                  className={classes}
-                  data-testid={`setup-runtime-tile-${spec.label}`}
-                  onClick={() => {
-                    if (!selectable) return;
-                    onToggleRuntime(spec.label);
-                  }}
-                  type="button"
-                  disabled={!selectable}
-                  aria-disabled={!selectable}
-                  aria-pressed={selected}
-                  title={
-                    installed
-                      ? detection?.version
-                        ? `${spec.label} — ${detection.version}`
-                        : spec.label
-                      : prereqsError
-                        ? `${spec.label} — detection failed, select if installed`
-                        : `${spec.label} — not installed`
-                  }
-                >
-                  {selected && (
-                    <span
-                      className="runtime-priority-badge"
-                      aria-label={`Priority ${priorityIdx + 1}`}
-                    >
-                      {priorityIdx + 1}
-                    </span>
-                  )}
-                  <div className="runtime-tile-head">
-                    <span
-                      className={`runtime-tile-status ${installed ? "installed" : ""}`}
-                      aria-hidden="true"
-                    />
-                    {spec.label}
-                  </div>
-                  <div className="runtime-tile-meta">
-                    {installed ? (
-                      detection?.version ? (
-                        detection.version
-                      ) : (
-                        "Installed"
-                      )
-                    ) : prereqsError ? (
-                      "Select if installed"
-                    ) : (
-                      <>
-                        Not installed{" · "}
-                        <a
-                          className="runtime-tile-install-link"
-                          href={spec.installUrl}
-                          target="_blank"
-                          rel="noopener noreferrer"
-                          onClick={(e) => e.stopPropagation()}
-                        >
-                          install
-                        </a>
-                      </>
-                    )}
-                  </div>
-                </button>
-              );
-            })}
-            {/* Run a local model — peer tile alongside the cloud CLIs.
-                Selecting it reveals the second-step picker (LocalLLMPicker)
-                below the grid. The dot stays grey because there's no single
-                binary to detect; the picker resolves which runtime is
-                actually running. */}
-            <button
-              type="button"
-              className={["runtime-tile", localModeOn ? "selected" : ""]
-                .filter(Boolean)
-                .join(" ")}
-              onClick={() => {
-                const next = !localModeOn;
-                setLocalModeOn(next);
-                if (!next) onSelectLocalProvider("");
-              }}
-              aria-pressed={localModeOn}
-              data-testid="onboarding-local-llm-toggle"
-              title="Run a local model on this machine"
-            >
-              <div className="runtime-tile-head">
-                <span className="runtime-tile-status" aria-hidden="true" />
-                Run a local model
-              </div>
-              <div className="runtime-tile-meta">
-                {localProvider
-                  ? (LOCAL_PROVIDER_LABELS.find((m) => m.kind === localProvider)
-                      ?.label ?? "selected")
-                  : "MLX-LM, Ollama, or Exo"}
-              </div>
-            </button>
-          </div>
-        )}
-
-        {localModeOn && (
-          <LocalLLMPicker
-            selected={localProvider}
-            onSelect={onSelectLocalProvider}
-          />
-        )}
-
-        {runtimePriority.length > 1 && (
-          <div className="runtime-priority-controls">
-            <p className="runtime-priority-title">Fallback order</p>
-            <p className="runtime-priority-hint">
-              Agents try these in order. Drop a local model into the chain so a
-              cloud quota hit falls through to your machine instead of
-              pay-as-you-go billing. Use the arrows to reorder.
-            </p>
-            {runtimePriority.map((label, idx) => (
-              <div key={label} className="runtime-priority-row">
-                <span className="runtime-priority-row-rank">#{idx + 1}</span>
-                <span className="runtime-priority-row-label">{label}</span>
-                <button
-                  type="button"
-                  className="runtime-priority-btn"
-                  onClick={() => onReorderRuntime(label, -1)}
-                  disabled={idx === 0}
-                  aria-label={`Move ${label} up`}
-                >
-                  ↑
-                </button>
-                <button
-                  type="button"
-                  className="runtime-priority-btn"
-                  onClick={() => onReorderRuntime(label, 1)}
-                  disabled={idx === runtimePriority.length - 1}
-                  aria-label={`Move ${label} down`}
-                >
-                  ↓
-                </button>
-                <button
-                  type="button"
-                  className="runtime-priority-btn"
-                  onClick={() => onToggleRuntime(label)}
-                  aria-label={`Remove ${label}`}
-                >
-                  ✕
-                </button>
-              </div>
-            ))}
-          </div>
-        )}
-
-        <div
-          style={{
-            marginTop: 16,
-            paddingTop: 16,
-            borderTop: "1px solid var(--border)",
-          }}
-        >
-          <p
-            style={{
-              fontSize: 13,
-              fontWeight: 600,
-              margin: "0 0 4px 0",
-              color: "var(--text)",
-            }}
-          >
-            API keys{" "}
-            {hasInstalledSelection ? "(optional fallback)" : "(required)"}
-          </p>
-          <p
-            style={{
-              fontSize: 12,
-              color: "var(--text-secondary)",
-              margin: "0 0 12px 0",
-            }}
-          >
-            {hasInstalledSelection
-              ? "Only used if every selected CLI fails. Leave blank to rely on the CLI login."
-              : "No installed CLI selected. Add at least one key so agents can reason."}
-          </p>
-          {API_KEY_FIELDS.map((field) => (
-            <ApiKeyRow
-              key={field.key}
-              field={field}
-              value={apiKeys[field.key] ?? ""}
-              onChange={(v) => onChangeApiKey(field.key, v)}
-            />
-          ))}
-        </div>
-      </div>
-
-      <div className="wizard-panel">
-        <p className="wizard-panel-title">Organizational memory</p>
-        <p
-          style={{
-            fontSize: 12,
-            color: "var(--text-secondary)",
-            margin: "-8px 0 12px 0",
-          }}
-        >
-          Where agents store shared context, relationships, and learnings across
-          sessions. You can change this later in Settings or via{" "}
-          <code>--memory-backend</code>.
-        </p>
-        <div className="runtime-grid">
-          {MEMORY_BACKEND_OPTIONS.map((opt) => (
-            <button
-              key={opt.value}
-              className={`runtime-tile ${memoryBackend === opt.value ? "selected" : ""}`}
-              onClick={() => onChangeMemoryBackend(opt.value)}
-              type="button"
-              title={opt.hint}
-            >
-              <div style={{ fontWeight: 600 }}>{opt.label}</div>
-              <div
-                style={{
-                  fontSize: 11,
-                  color: "var(--text-tertiary)",
-                  marginTop: 4,
-                  fontWeight: 400,
-                }}
-              >
-                {opt.hint}
-              </div>
-            </button>
-          ))}
-        </div>
-
-        {gbrainSelected && (
-          <div className="wiz-backend-keys">
-            <p className="wiz-backend-keys-title">GBrain keys</p>
-            <p className="wiz-backend-keys-hint">
-              GBrain uses OpenAI for embeddings (required) and optionally
-              Anthropic for reasoning.
-            </p>
-            <div className="form-group">
-              <label className="label" htmlFor="wiz-gbrain-openai">
-                OpenAI API key <span style={{ color: "var(--red)" }}>*</span>
-              </label>
-              <input
-                className="input"
-                id="wiz-gbrain-openai"
-                type="password"
-                placeholder="sk-..."
-                value={gbrainOpenAIKey}
-                onChange={(e) => onChangeGBrainOpenAIKey(e.target.value)}
-                autoComplete="off"
-              />
-              {gbrainOpenAIMissing && (
-                <p style={{ color: "var(--red)", fontSize: 11, marginTop: 4 }}>
-                  Required: GBrain can&apos;t create embeddings without an
-                  OpenAI key.
-                </p>
-              )}
-            </div>
-            <div className="form-group" style={{ marginBottom: 0 }}>
-              <label className="label" htmlFor="wiz-gbrain-anthropic">
-                Anthropic API key{" "}
-                <span style={{ fontSize: 11, color: "var(--text-tertiary)" }}>
-                  (optional)
-                </span>
-              </label>
-              <input
-                className="input"
-                id="wiz-gbrain-anthropic"
-                type="password"
-                placeholder="sk-ant-..."
-                value={gbrainAnthropicKey}
-                onChange={(e) => onChangeGBrainAnthropicKey(e.target.value)}
-                autoComplete="off"
-              />
-            </div>
-          </div>
-        )}
-      </div>
-
-      {memoryBackend === "nex" && (
-        // Only show the Nex API key panel when the chosen memory backend
-        // actually needs it. Team wiki, GBrain, and "None" don't talk to
-        // Nex's hosted memory — surfacing the input would suggest the
-        // user has a missing piece when they don't.
-        <div className="wizard-panel" data-testid="wizard-nex-api-key-panel">
-          <p className="wizard-panel-title">Nex API key</p>
-          <p
-            style={{
-              fontSize: 12,
-              color: "var(--text-secondary)",
-              margin: "-8px 0 12px 0",
-            }}
-          >
-            Unlocks the hosted memory graph plus managed integrations (HubSpot,
-            Slack, Gmail, Calendar, …) so agents can read your tools without you
-            wiring each one up. You can skip this and paste later from Settings.
-            Don&apos;t have one? Sign up on the Identity step above.
-          </p>
-          <div className="form-group" style={{ marginBottom: 0 }}>
-            <label className="label" htmlFor="wiz-nex-api-key">
-              Nex API key{" "}
-              <span style={{ fontSize: 11, color: "var(--text-tertiary)" }}>
-                (optional, paste if you have one)
-              </span>
-            </label>
-            <input
-              className="input"
-              id="wiz-nex-api-key"
-              type="password"
-              placeholder="nex-..."
-              value={nexApiKey}
-              onChange={(e) => onChangeNexApiKey(e.target.value)}
-              autoComplete="off"
-            />
-          </div>
-        </div>
-      )}
-
-      <div className="wizard-nav">
-        <button className="btn btn-ghost" onClick={onBack} type="button">
-          Back
-        </button>
-        <button
-          className="btn btn-primary"
-          onClick={onNext}
-          disabled={!canContinue}
-          type="button"
-        >
-          {ONBOARDING_COPY.step2_cta}
-          <ArrowIcon />
-          <EnterHint />
-        </button>
-      </div>
-    </div>
-  );
-}
-
-/* ─── Step 6: First Task ─── */
-
-interface TaskStepProps {
-  taskTemplates: TaskTemplate[];
-  selectedTaskTemplate: string | null;
-  onSelectTaskTemplate: (id: string | null) => void;
-  taskText: string;
-  onChangeTaskText: (v: string) => void;
-  onNext: () => void;
-  onSkip: () => void;
-  onBack: () => void;
-  submitting: boolean;
-}
-
-function TaskStep({
-  taskTemplates,
-  selectedTaskTemplate,
-  onSelectTaskTemplate,
-  taskText,
-  onChangeTaskText,
-  onNext,
-  onSkip,
-  onBack,
-  submitting,
-}: TaskStepProps) {
-  return (
-    <div className="wizard-step">
-      <div className="wizard-hero">
-        <h1 className="wizard-headline" style={{ fontSize: 28 }}>
-          {ONBOARDING_COPY.step3_title}
-        </h1>
-        {taskTemplates.length > 0 && (
-          <p className="wizard-subhead">
-            Type your own first task, or pick from the blueprint&apos;s
-            suggested sequence below.
-          </p>
-        )}
-      </div>
-
-      <div>
-        <textarea
-          className="task-textarea task-textarea-primary"
-          id="wiz-task-input"
-          placeholder={ONBOARDING_COPY.step3_placeholder}
-          value={taskText}
-          onChange={(e) => onChangeTaskText(e.target.value)}
-        />
-        <p className="task-textarea-hint">
-          <Kbd size="sm">↵</Kbd> new line · <Kbd size="sm">{MOD_KEY}</Kbd>
-          <Kbd size="sm">↵</Kbd> review setup
-        </p>
-      </div>
-
-      {taskTemplates.length > 0 && (
-        <div className="task-suggestions">
-          <p className="task-suggestions-label">
-            Suggested sequence for this blueprint
-          </p>
-          <div className="task-suggestions-list">
-            {taskTemplates.map((t, idx) => {
-              const isSelected = selectedTaskTemplate === t.id;
-              return (
-                <button
-                  key={t.id}
-                  className={`task-suggestion ${isSelected ? "selected" : ""}`}
-                  onClick={() => {
-                    const nextId = isSelected ? null : t.id;
-                    onSelectTaskTemplate(nextId);
-                    if (nextId) {
-                      onChangeTaskText(t.prompt ?? t.name);
-                    }
-                  }}
-                  type="button"
-                >
-                  <span className="task-suggestion-num">{idx + 1}</span>
-                  <span className="task-suggestion-name">{t.name}</span>
-                </button>
-              );
-            })}
-          </div>
-        </div>
-      )}
-
-      <div className="wizard-nav">
-        <button className="btn btn-ghost" onClick={onBack} type="button">
-          Back
-        </button>
-        <div className="wizard-nav-right">
-          <button
-            className="task-skip"
-            onClick={onSkip}
-            disabled={submitting}
-            type="button"
-          >
-            {ONBOARDING_COPY.step3_skip}
-          </button>
-          <button className="btn btn-primary" onClick={onNext} type="button">
-            Review setup
-            <ArrowIcon />
-            <EnterHint modifier={MOD_KEY} />
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
-/* ─── Step 7: Readiness Summary ─── */
-
-// ReadinessStatus mirrors the TUI's three-state readiness color mapping
-// (see internal/tui/init_flow.go readinessStatusColor): 'ready' = green
-// check, 'next' = blue warning (follow-up needed), 'missing' = red.
-
-interface ReadyStepProps {
-  checks: ReadinessCheck[];
-  taskText: string;
-  submitting: boolean;
-  submitError: string;
-  onSkip: () => void;
-  onSubmit: () => void;
-  onBack: () => void;
-}
-
-// ReadyStep is the six-item final review matching the TUI's InitDone
-// readinessChecks() view. It's honest: a missing Nex key is not papered
-// over, and GBrain+no-OpenAI-key would show a red "missing" row (though
-// the Setup step blocks continuing in that case, so users shouldn't get
-// here with it).
-function ReadyStep({
-  checks,
-  taskText,
-  submitting,
-  submitError,
-  onSkip,
-  onSubmit,
-  onBack,
-}: ReadyStepProps) {
-  return (
-    <div className="wizard-step">
-      <div className="wizard-hero">
-        <h1 className="wizard-headline" style={{ fontSize: 28 }}>
-          You&apos;re set
-        </h1>
-        <p className="wizard-subhead">
-          Here&apos;s what&apos;s configured. Anything with a{" "}
-          <span className="readiness-glyph-inline missing">!</span> or{" "}
-          <span className="readiness-glyph-inline next">—</span> can be fixed
-          later from Settings.
-        </p>
-      </div>
-
-      <div className="wizard-panel readiness-panel">
-        <ul className="readiness-list">
-          {checks.map((check) => (
-            <li key={check.label} className="readiness-item">
-              <span
-                className={`readiness-glyph ${check.status}`}
-                aria-hidden="true"
-              >
-                {check.status === "ready"
-                  ? "✓"
-                  : check.status === "next"
-                    ? "—"
-                    : "!"}
-              </span>
-              <div className="readiness-body">
-                <div className="readiness-label">{check.label}</div>
-                <div className="readiness-detail">{check.detail}</div>
-              </div>
-            </li>
-          ))}
-        </ul>
-      </div>
-
-      {submitError && (
-        <div
-          role="alert"
-          data-testid="onboarding-submit-error"
-          style={{
-            fontSize: 13,
-            color: "var(--danger-500, #c33)",
-            padding: "12px 14px",
-            background: "var(--danger-50, #fee)",
-            border: "1px solid var(--danger-200, #fcc)",
-            borderRadius: 6,
-          }}
-        >
-          <strong>Could not start the office:</strong> {submitError}. Check that
-          your CLI runtime is installed and try again, or go back to adjust your
-          setup.
-        </div>
-      )}
-
-      <div className="wizard-nav">
-        <button className="btn btn-ghost" onClick={onBack} type="button">
-          Back
-        </button>
-        <div className="wizard-nav-right">
-          <button
-            className="btn btn-primary"
-            data-testid="onboarding-submit-button"
-            onClick={taskText.trim().length === 0 ? onSkip : onSubmit}
-            disabled={submitting}
-            type="button"
-          >
-            {submitting
-              ? "Starting..."
-              : submitError
-                ? "Retry"
-                : ONBOARDING_COPY.step3_cta}
-            {!submitting && taskText.trim().length > 0 && <EnterHint />}
-          </button>
-        </div>
-      </div>
-    </div>
-  );
-}
-
+// Step components extracted to wizard/Step1Welcome.tsx through Step7Ready.tsx.
+// runtimeIsReady + detectedBinary moved to wizard/runtime-helpers.ts.
 /* ═══════════════════════════════════════════
    Main Wizard
    ═══════════════════════════════════════════ */
